% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelValidation.R
\name{print.ValidationResult}
\alias{print.ValidationResult}
\title{Print method for ValidationResult objects}
\usage{
\method{print}{ValidationResult}(x, ...)
}
\arguments{
\item{x}{A ValidationResult object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for ValidationResult objects
}
