% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HadesResultsModel.R
\docType{package}
\name{HadesResultsModel-package}
\alias{HadesResultsModel}
\alias{HadesResultsModel-package}
\title{HadesResultsModel: HADES Results Model Definitions and Utilities}
\description{
Lightweight R package for working with OHDSI HADES results model resources, including module YAML definitions, release manifests, JSON schemas, and migration and DDL generation utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/HadesResultsModel/}
  \item \url{https://github.com/OHDSI/HadesResultsModel}
  \item Report bugs at \url{https://github.com/OHDSI/HadesResultsModel/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Martijn Schuemie \email{schuemie@ohdsi.org}
}

}
\keyword{internal}
